backend_geruest/wattwert_only_mtl/requirements.txt
torch
torchvision
albumentations
evaluate
tqdm
opencv-python
matplotlib
pyyaml
numpy
tensorboard
scipy
scikit-image
scikit-learn
pillow